refactor: split runtime fallback auto retry facade
