+++ src/hooks/runtime-fallback/auto-retry-dispatch.ts
@@ -1,0 +1,183 @@
+import type { HookDeps } from "./types"
+import { HOOK_NAME } from "./constants"
+import { log } from "../../shared/logger"
+import { getSessionAgent, resolveRegisteredAgentName } from "../../features/claude-code-session-state"
+import { buildRetryModelPayload } from "./retry-model-payload"
+import { getLastUserRetryPayload } from "./last-user-retry-parts"
+import { createInternalAgentContinuationTextPart } from "../../shared/internal-initiator-marker"
+import {
+  dispatchInternalPrompt,
+  isInternalPromptDispatchAccepted,
+} from "../shared/prompt-async-gate"
+import { isAmbiguousPostDispatchPromptFailure } from "../../shared/prompt-failure-classifier"
+import { resolveOriginalUserRetryMetadata } from "./auto-retry-metadata"
+
+export function createAutoRetryDispatcher(
+  deps: HookDeps,
+  scheduleSessionFallbackTimeout: (sessionID: string, resolvedAgent?: string) => void,
+  clearSessionFallbackTimeout: (sessionID: string) => void,
+) {
+  const {
+    ctx,
+    sessionStates,
+    sessionRetryInFlight,
+    sessionAwaitingFallbackResult,
+    pluginConfig,
+  } = deps
+
+  return async (
+    sessionID: string,
+    newModel: string,
+    resolvedAgent: string | undefined,
+    source: string,
+  ): Promise<void> => {
+    if (sessionRetryInFlight.has(sessionID)) {
+      log(`[${HOOK_NAME}] Retry already in flight, skipping (${source})`, { sessionID })
+      return
+    }
+
+    const agentSettings = resolvedAgent
+      ? pluginConfig?.agents?.[resolvedAgent as keyof typeof pluginConfig.agents]
+      : undefined
+    const retryModelPayload = buildRetryModelPayload(newModel, agentSettings ? {
+      variant: agentSettings.variant,
+      reasoningEffort: agentSettings.reasoningEffort,
+    } : undefined)
+    if (!retryModelPayload) {
+      log(`[${HOOK_NAME}] Invalid model format (missing provider prefix): ${newModel}`)
+      const state = sessionStates.get(sessionID)
+      if (state?.pendingFallbackModel) {
+        state.pendingFallbackModel = undefined
+      }
+      if (state) {
+        state.pendingFallbackPromptMayHaveBeenAccepted = false
+      }
+      return
+    }
+
+    const hadAwaitingFallbackResult = sessionAwaitingFallbackResult.has(sessionID)
+    const previousPendingFallbackModel = sessionStates.get(sessionID)?.pendingFallbackModel
+    const previousPendingFallbackPromptMayHaveBeenAccepted = sessionStates.get(sessionID)?.pendingFallbackPromptMayHaveBeenAccepted
+    sessionRetryInFlight.add(sessionID)
+    let retryDispatched = false
+    let retryMayHaveBeenAccepted = false
+    try {
+      const messagesResp = await ctx.client.session.messages({
+        path: { id: sessionID },
+        query: { directory: ctx.directory },
+      })
+      const retryPayload = getLastUserRetryPayload(messagesResp, sessionID)
+      const originalRetryMetadata = resolveOriginalUserRetryMetadata(messagesResp)
+      const fetchedParts = originalRetryMetadata.parts.length > 0
+        ? originalRetryMetadata.parts
+        : retryPayload.retryParts
+      const usingFetchedUserParts = originalRetryMetadata.parts.length > 0
+      const retryParts =
+        fetchedParts.length > 0
+          ? fetchedParts
+          : (() => {
+              log(
+                `[${HOOK_NAME}] No user message parts found for auto-retry (${source}); using synthetic continuation`,
+                {
+                  sessionID,
+                  hint: "This can occur when the working directory contains .git and messages are not yet persisted",
+                },
+              )
+              // Mark the retry as internally initiated so continuation hooks
+              // do not render a user-visible bare "continue" turn (#4085).
+              return [createInternalAgentContinuationTextPart("continue")]
+            })()
+      const retryMessageID = usingFetchedUserParts ? originalRetryMetadata.messageID : undefined
+      log(`[${HOOK_NAME}] Auto-retrying with fallback model (${source})`, {
+        sessionID,
+        model: newModel,
+      })
+
+      const retryAgent = resolvedAgent ?? getSessionAgent(sessionID)
+      const launchAgent = resolveRegisteredAgentName(retryAgent)
+      if (!hadAwaitingFallbackResult) {
+        sessionAwaitingFallbackResult.add(sessionID)
+        scheduleSessionFallbackTimeout(sessionID, retryAgent)
+      }
+
+      const promptResult = await dispatchInternalPrompt({
+        mode: "async",
+        client: ctx.client,
+        sessionID,
+        source: `runtime-fallback:${source}`,
+        settleMs: 0,
+        queueBehavior: "defer",
+        input: {
+          path: { id: sessionID },
+          body: {
+            ...(launchAgent ? { agent: launchAgent } : {}),
+            ...retryModelPayload,
+            ...(retryPayload.system ? { system: retryPayload.system } : {}),
+            ...(retryPayload.tools ? { tools: retryPayload.tools } : {}),
+            ...(retryMessageID ? { messageID: retryMessageID } : {}),
+            parts: retryParts,
+          },
+          query: { directory: ctx.directory },
+        },
+      })
+      if (promptResult.status === "failed") {
+        if (isAmbiguousPostDispatchPromptFailure(promptResult)) {
+          retryMayHaveBeenAccepted = true
+          log(`[${HOOK_NAME}] Auto-retry prompt failed after dispatch may have been accepted (${source}); preserving fallback state`, {
+            sessionID,
+            error: String(promptResult.error),
+          })
+        }
+        throw promptResult.error
+      }
+      if (!isInternalPromptDispatchAccepted(promptResult)) {
+        log(`[${HOOK_NAME}] Auto-retry skipped by promptAsync gate (${source})`, {
+          sessionID,
+          status: promptResult.status,
+        })
+        return
+      }
+      sessionAwaitingFallbackResult.add(sessionID)
+      if (hadAwaitingFallbackResult) {
+        scheduleSessionFallbackTimeout(sessionID, retryAgent)
+      }
+      const state = sessionStates.get(sessionID)
+      if (state) {
+        state.pendingFallbackPromptMayHaveBeenAccepted = false
+      }
+      retryDispatched = true
+    } catch (retryError) {
+      if (!(retryError instanceof Error)) {
+        log(`[${HOOK_NAME}] Auto-retry failed (${source})`, { sessionID, error: String(retryError) })
+        return
+      }
+      log(`[${HOOK_NAME}] Auto-retry failed (${source})`, { sessionID, error: String(retryError) })
+    } finally {
+      sessionRetryInFlight.delete(sessionID)
+      if (retryMayHaveBeenAccepted) {
+        const state = sessionStates.get(sessionID)
+        if (state) {
+          state.pendingFallbackPromptMayHaveBeenAccepted = true
+        }
+      }
+      if (!retryDispatched && !retryMayHaveBeenAccepted) {
+        if (hadAwaitingFallbackResult) {
+          sessionAwaitingFallbackResult.add(sessionID)
+        } else {
+          sessionAwaitingFallbackResult.delete(sessionID)
+          clearSessionFallbackTimeout(sessionID)
+        }
+        const state = sessionStates.get(sessionID)
+        if (state) {
+          if (hadAwaitingFallbackResult) {
+            state.pendingFallbackModel = previousPendingFallbackModel
+            state.pendingFallbackPromptMayHaveBeenAccepted = previousPendingFallbackPromptMayHaveBeenAccepted
+          } else if (state.pendingFallbackModel) {
+            state.pendingFallbackModel = undefined
+            state.pendingFallbackPromptMayHaveBeenAccepted = false
+          }
+        }
+      }
+    }
+  }
+}
